Allowed text renderer to run on second adaptor. Split postprocessing into 3D scene and fianl image effects. Added compost post process effect. Added ensureMsgf and checkMsgf

+++ GraphicsEngine/GraphicsEngine/Core/Utils/MovingAverage.h
@@ -20,7 +20,7 @@ private:
 	float * Array;
 	float Sum;
 	float Average;
-	unsigned short Index;
+	int Index;
 	bool FilterComplete;
 	void init();
 };
